# docs/04-api/endpoints/openapi-spec.yaml
openapi: 3.1.0
info:
  title: ToolboxAI Roblox Environment
  description: >-
    AI-Powered Educational Roblox Environment - Generate immersive educational
    content with AI agents
  version: 1.0.0
paths:
...
  /reports/{report_id}:
    get:
      tags:
        - Reports
      summary: Get Report Details
      description: Get detailed information about a specific report.
      operationId: get_report_details_reports__report_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: integer
            title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Report Details Reports  Report Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer